requirements.txt
Click==7.0
python-nmap==0.6.1
